moving wishlist features to other components - test
