+++ IT-Wizards-UI/src/components/ItemDisplay.jsx
@@ -3,13 +3,17 @@ import { getItems } from '../services/viewItemsService';
 import cauldron from '../assets/images/cauldron.png';
 import { Link } from 'react-router-dom';
 import { CartContext } from './CartContext';
+import { addItemToWishlist, removeItemFromWishlist, getWishlist } from '../services/wishlistService';
 
 const ItemDisplay = ({ searchTerm }) => {
   const cart = useContext(CartContext);
   const [items, setItems] = useState([]);
+  const [wishlist, setWishlist] = useState([]);
+  const userId = '${userId}';
 
   useEffect(() => {
     fetchItems();
+    fetchWishlist();
   }, []);
 
   const fetchItems = async () => {
@@ -20,6 +24,42 @@ const ItemDisplay = ({ searchTerm }) => {
       console.error('Failed to fetch data', error);
     }
   };
+
+  const fetchWishlist = async () => {
+    try {
+        const data = await getWishlist(userId);
+        setWishlist(data);
+    }catch (error) {
+        console.error('Error fetching wishlist', error);
+    }
+};
+
+  const inWishlist = (itemId) => {
+    for (let i = 0; i<wishlist.length; i++) {
+        if (wishlist[i].item_id === itemId) {
+            return true;
+        }
+    }
+    return false;
+};
+
+const  addToWishlist= async (itemId) => {
+  try {
+      await addItemToWishlist(userId, itemId);
+      fetchWishlist();
+  } catch (error) {
+      console.error('Error adding item from wishlist', error);
+  }
+};
+
+const removeFromWishlist = async (itemId) => {
+  try { 
+      await removeItemFromWishlist(userId, itemId);
+      fetchWishlist();
+  } catch (error) {
+      console.error('Error removing item from wishlist', error);
+  }
+};
 
   const itemsArr = [];
 
@@ -63,6 +103,18 @@ const ItemDisplay = ({ searchTerm }) => {
                           <div className="flex items-center justify-center">
                             ${(Math.round(item.price * 100) / 100).toFixed(2)}
                           </div>
+                          {inWishlist(item.id) ? (
+                            <button
+                            className="bg-red-600 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-full w-full mt-3"
+                            onClick={()=> removeFromWishlist(item.id)}
+                            > Remove From Wishlist </button>
+                          ) : (
+                            <button
+                            className="bg-green-600 hover:bg-green-700 text-white font-bold py-2 px-4 rounded-full w-full mt-3"
+                            onClick={()=> addToWishlist(item.id)}
+                            > Add To Wishlist</button>
+                          )
+                        }
 
                           {cart.getItemQuantity(item.id) > 0 ? (
                             <div className="">
